Refactoring some pages into their own folders
