+++ src/components/page/Footer/styles.js
@@ -1,0 +1,7 @@
+import { Container } from "@mui/material";
+import { styled } from "@mui/system";
+
+export const FooterContainer = styled(Container)(({ theme }) => ({
+    padding: "20px",
+    backgroundColor: theme.palette.landing
+}));
